# PERF: improves performance in GroupBy.cumcount