@jreback 2015-09-10
this would prob close #7569 
and partial on #5755 

pls add tests and/or advise.

ty