@behzadnouri 2015-09-10
well i am not adding a new feature or closing a bug; so the tests already there should be fine