@hayd 2015-09-10
I think this crept in via #7910 / @mortada adding support for passing lists to nth.