@jreback 2015-09-11
I think this should be changed as @behzadnouri suggest. Code is MUCH simpler. and everything would be consistent. I think what ``nth`` does now is a little odd (and IIRC is differently if you ``.apply`` it rather than directly use ``.nth``)