@hayd 2015-09-12
nth was supposed to be of the "filtering" type rather than aggregating type. Hence the v0.16.2 behaviour.