@jreback 2015-09-13
moving to next version. This needs to not change the API. The result index can be set differently at the end.