@behzadnouri 2015-09-13
api is the function signature which does not change. 

if you mean the behaviour, it is inconsistent between series and data-frame, and _even within series if there are nulls_; and seems to be a bug, not by design. so at some point you need to make series behave like frames or the other way around.

```python
>>> df
   A  B
0  a  0
1  b  1
>>> df.groupby('A').nth(0)
   B
A   
a  0
b  1
>>> df.groupby('A')['B'].nth(0)
0    0
1    1
Name: B, dtype: int64

>>> ts
0     0
1   NaN
dtype: float64
>>> ts.groupby(Series(['a', 'b'])).nth(0, dropna=True)
a   NaN
b   NaN
dtype: float64
```