@hayd 2015-09-13
Series should behave like DataFrame (and filter rather than agg). I was *sure* these worked the same on Series or it did behind a flag (which was not able to be the default due to old behaviour).