@jorisvandenbossche 2015-09-13
@hayd I think we are not using the same terminology :-)

Currently, nth DataFrame behaviour is *aggregating/reducing* (so setting the grouping keys as the index), while Series behaviour is *filtering* (keeping the original index labels) (so the other way around than I understand from your comment)

Originally `nth` on a DataFrame was filtering when you added it (https://github.com/pydata/pandas/pull/6569), but it was changed to reducing even before it was included in a release by @jreback  (https://github.com/pydata/pandas/pull/7044). Back then I asked specifically about this (https://github.com/pydata/pandas/pull/7044#issuecomment-42215735), but @jreback gave some reasons to make it reducing. Only, in that PR, I think it was an oversight of us that it only made DataFrame reducing, and left Series behaviour as filtering. 
Also in the docs, `nth` is mentioned in the reducing methods explicitely: the 'note box at the end of this section: http://pandas.pydata.org/pandas-docs/stable/groupby.html#aggregation

By the introduction of the possibility to pass lists to get multiple values, it complicated this a bit, as a real aggregating method will only return one value per group. Further, with the introduction of `dropna` kwarg, you cannot really call it filtering anymore when using that, since it will not return original rows always.