@jreback 2015-09-13
IIRC (and has been a while), we want to make ``nth(0,dropna=True) == .first()`` and ``nth(0,dropna=True) == .last()``