@jorisvandenbossche 2015-09-13
@jreback and that is true for a DataFrame, but not Series (so this PR makes that analogy more consistent)