@jorisvandenbossche 2015-09-13
Now, for a solution. I think we just have to make a choice ..

1) Leave it as is (DataFrame as reducing, Series as filtering, it has been like this since 0.14)
2) Have all as reducing (so change Series)
3) Have all as filtering (and change DataFrame behaviour)

If we change something, I would go for 2), as the filtering behaviour is easy to get by using `as_index=False` (while for the other way around you have to do `set_index()`) + then it is more consistent with `first/last`