@jreback 2015-09-13
ahh, ok this was a consistency-fix partially then. I would agree with 2).