@jorisvandenbossche 2015-09-13
Some other inconsistencies:

* `as_index=False` is not working when using dropna:

  ```
In [38]: df = pd.DataFrame({'a':list('ABABAB'), 'b':[1,2,3,4,5,6]})

In [39]: df2 = df.copy()

In [40]: df2.iloc[1,1] = np.nan

In [41]: df2.groupby('a').nth(0, dropna='any')
Out[41]:
   b
a
A  1
B  4

In [42]: df2.groupby('a', as_index=False).nth(0, dropna='any')
Out[42]:
   b
a
A  1
B  4

In [43]: df2.groupby('a', as_index=False).nth(0)
Out[43]:
   a   b
0  A   1
1  B NaN
  ```

* the resulting index when using `as_index=False` is different between `nth` and `first`:

```
In [45]: df.index = df.index + 10

In [46]: df.groupby('a').nth(0)
Out[46]:
   b
a
A  1
B  2

In [47]: df.groupby('a').first()
Out[47]:
   b
a
A  1
B  2

In [48]: df.groupby('a', as_index=False).first()
Out[48]:
   a  b
0  A  1
1  B  2

In [50]: df.groupby('a', as_index=False).nth(0)
Out[50]:
    a  b
10  A  1
11  B  2
```