@jreback 2015-09-13
these are already noted in #5755 (though not so explicity)