@jorisvandenbossche 2015-09-13
But here it is possibly on purpose, as for `nth` this is a way to get filtering behaviour (keep original index), while reducing methods give you just a default 0, 1, .., n index when using `as_index=False` as you never can have the original index. That is the problem with `nth` being some mixture of a reducing and filtering method.