@jreback 2015-11-25
@behzadnouri can you rebase / update

further need to clarify any API changes / fixes here