@hayd 2015-11-25
@jorisvandenbossche I don't understand why the same argument couldn't apply to .head, isn't .nth is the same in whether it's filtering/reducing. :( Or is this claim due to dropping NaN, the *really* slow path inside nth.

IME the nth, index behaviour is rarely what you want, you just want the nth line regardless of NaN. Maybe we should have an iloc method/accessor on groupby that would be more explicit (and always fast) and do just that.