@jreback 2015-11-25
@hayd if you wouldn't mind putting up some examples of what you think we should be doing for the various filter/reducers: ``.head/nth/first/iloc`` (and I agree ``.iloc`` would be a nice clear accessor, though obviously wouldn't skip ``nan`, we did in fact had ``.irow``, but deprecated recently)