@hayd 2015-11-26
yeah, iirc it was deprecated cos it was slow... it seems that the nth path (which skips nans) is similarly slow. Do people use this feature/is it useful? I guess if we had iloc, as the fast path, we (I) can just ignore nth as an R-throwback. :p

I will put together some examples/ideas next week.