@jreback 2016-01-24
ok, so who wants to update, propose something here 2) seemed to be the most useful way to go, IOW, fix Series to be reducing (as DataFrame is now), and easy to get filtering with ``as_index=False``

cc @behzadnouri 
@hayd 
@jorisvandenbossche